requirements.txt
asgiref==3.4.1
astroid==2.8.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-environ==0.7.0
django-extensions==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.3.0
psycopg2==2.9
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.12.1
